server/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
faiss-cpu
sentence-transformers
pymupdf
torch
python-multipart